LayoutTests/fast/dom/Window/Location/window-override-window-using-defineGetter.html

<html>
<head>
    <script src="../../../../resources/js-test-pre.js"></script>
    <script src="resources/location-tests-functions.js"></script>
</head>
<body>
<script>
    shouldThrowErrorName(function() {
        window.__defineGetter__("window", () => ({ location: "haxored" }));
    }, "TypeError");

    var result = normalizeURL(String(window.location));
    var correctValue = normalizeURL(document.URL);
    shouldBe("result", "correctValue");
</script>
<script src="../../../../resources/js-test-post.js"></script>
</body>
</html>
