LayoutTests/fast/dom/Window/Location/window-override-window-using-defineGetter-expected.txt

PASS function () {
        window.__defineGetter__("window", () => ({ location: "haxored" }));
    } threw exception TypeError: Attempting to change configurable attribute of unconfigurable property..
PASS result is correctValue
PASS successfullyParsed is true

TEST COMPLETE
